Adding initial conversion of TCPRS to a plugin
TCPRS is TCP traffic analyzer that specializes in the detection
and classification of retransmission and network reordering events.

The following forms of events are available in the TCPRS analyzer:

* Dead connection detection
* TCP option detection
* Retransmission detection and classification
* Limited Transmit and Fast Recovery detection
* Network reordering detection and classification
* RTT and initial RTO measurements

Included with the analyzer is a collection of 103 test cases that
are used for iterative design and refinement of the analyzer. Each
test case is used to verify a specific function of the analyzer or
general classification of events.

Signed-off-by: James Swaro <james.swaro@gmail.com>